Version 1.0

+++ priv/repo/migrations/20211111112254_add_rubins.exs
@@ -3,7 +3,7 @@ defmodule Makoto.Repo.Migrations.AddRubins do
 
   def change do
     alter table(:users) do
-      add(:rubins, :integer, default: 0)
+      add(:rubins, :float, default: 0)
     end
   end
 end
